+++ app/src/main/java/com/example/kakaoimglibrary/common/Constants.kt
@@ -1,4 +1,4 @@
-package com.example.kakaoimglibrary
+package com.example.kakaoimglibrary.common
 
 class Constants {
     companion object{
