+++ app/src/main/java/com/example/kakaoimglibrary/common/EntryType.kt
@@ -1,4 +1,4 @@
-package com.example.kakaoimglibrary.main
+package com.example.kakaoimglibrary.common
 
 enum class EntryType {
     ADD, EDIT, REMOVE;
